// src/enums/credit-check.ts
export enum CreditCheckStatus {
  WAITING_FOR_PAYMENT = 'WAITING_FOR_PAYMENT',
  WAITING_FOR_CONFIRMATION = 'WAITING_FOR_CONFIRMATION',
  IN_PROGRESS = 'IN_PROGRESS',
  DONE_CHECKING = 'DONE_CHECKING',
  EXPIRED = 'EXPIRED',
}

export enum CreditCheckFileCategory {
  ATTACHMENT = 'ATTACHMENT',
  PAYMENT_PROOF = 'PAYMENT_PROOF',
  VENDOR = 'VENDOR',
}